Export Rust build environment for use by other buildpacks
See #9.  The theory here is that we might have a Ruby application that
uses a gem implemented in Rust, and in order to compile the gem, the
Ruby buildpack needs access to the Rust compiler.

To support this, we need to create an `export` file in our buildpack
directory containing all the environment variables needed to run `cargo`
and `rustc`.

Note that this still has some limitations: You almost certainly need to
have a valid `Cargo.toml` in your project for the buildpack to detect
and build before you can run the other buildpack that needs Rust.